Always use `bytes` instead of `six.binary_type`
`bytes` was introduced in Python 2.6 as an alias for `str`.
`six.binary_type` is equivalent to it, but it makes sense to use the native
type instead of a compatibilty libray type when there's no difference in
them.

`bytes` was already being used in most places, this commit replaces the
three remaining uses.